Move Extensions.UnityEngine assembly into Extensions.UnityEngine namespace
