+++ Assets/Scripts/Animation/AnimationKeyFrame.cs
@@ -1,5 +1,5 @@
 using PAC.DataStructures;
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Json;
 using PAC.Json.JsonConverters;
