+++ Assets/Scripts/Clipboard/Clipboard.cs
@@ -1,7 +1,7 @@
 using PAC.Animation;
 using PAC.Colour.Compositing;
 using PAC.Drawing;
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Files;
 using PAC.Geometry;
 using PAC.ImageEditing;
