+++ Assets/Scripts/Colour/Compositing/AlphaCompositing.cs
@@ -1,4 +1,4 @@
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 
 using UnityEngine;
