+++ Assets/Scripts/Colour/Compositing/BlendMode.cs
@@ -2,7 +2,7 @@ using System;
 using System.Linq;
 using System.Reflection;
 
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Geometry;
 using PAC.Json;
