+++ Assets/Scripts/Files/FileManager.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.IO;
 
 using PAC.Animation;
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Input;
 using PAC.Layers;
 using PAC.UI;
