+++ Assets/Scripts/Geometry/Extensions/IntVector2Extensions.cs
@@ -1,4 +1,4 @@
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Geometry.Extensions;
 
 using UnityEngine;
