+++ Assets/Scripts/Geometry/Shapes/Extensions/IShapeExtensions.cs
@@ -1,4 +1,4 @@
-﻿using PAC.Extensions;
+﻿using PAC.Extensions.UnityEngine;
 using PAC.Geometry.Extensions;
 using PAC.Geometry.Shapes.Interfaces;
