+++ Assets/Scripts/Image Editing/ImageEdit.cs
@@ -2,6 +2,7 @@ using System;
 
 using PAC.Exceptions;
 using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Extensions;
