+++ Assets/Scripts/Image Editing/Outline.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 
 using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Geometry;
 using PAC.Geometry.Extensions;
