+++ Assets/Scripts/Layers/Layer.cs
@@ -5,7 +5,7 @@ using System.Runtime.Serialization;
 using PAC.Animation;
 using PAC.Colour.Compositing;
 using PAC.DataStructures;
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Json;
