+++ Assets/Scripts/Layers/NormalLayer.cs
@@ -4,7 +4,7 @@ using System.Runtime.Serialization;
 using PAC.Animation;
 using PAC.Colour.Compositing;
 using PAC.DataStructures;
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.ImageEditing;
