+++ Assets/Scripts/Patterns/Extensions/IPattern2DExtensions.cs
@@ -1,6 +1,7 @@
 using System;
 
 using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Geometry;
 
 using UnityEngine;
