+++ Assets/Scripts/UI/UITileIcon.cs
@@ -1,4 +1,4 @@
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Files;
 
 using UnityEngine;
