+++ Assets/Tests/Colour/Compositing/BlendMode_Tests.cs
@@ -5,6 +5,7 @@ using NUnit.Framework;
 using PAC.Colour;
 using PAC.Colour.Compositing;
 using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 using PAC.Json;
 
 using UnityEngine;
