+++ Assets/Tests/Extensions/UnityEngine/ColorExtensions_Tests.cs
@@ -1,6 +1,6 @@
 using NUnit.Framework;
 
-using PAC.Extensions;
+using PAC.Extensions.UnityEngine;
 
 using UnityEngine;
